+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Users can now save login data to Password Vault automatically after login to an 
 * Visual enhancements within the dialog to change between light and dark mode.
 * Added new backups which can store settings and quality gates.
 * Added new screen to configure and create backups.
+* Removed dialog to create backups, since it has been replaced by a new activity.
 
 <br/>